Translate cash from int to double

+++ atm_core/credit_department/interest_rate_generator.cpp
@@ -2,8 +2,8 @@
 
 double InterestRateGenerator::kMaximalRate_ = 14.0;
 double InterestRateGenerator::kMinimalRate_ = 5.0;
-int InterestRateGenerator::kMinimalCash_ = 1000;
-int InterestRateGenerator::kMaximalCash_ = 20000;
+double InterestRateGenerator::kMinimalCash_ = 1000.0;
+double InterestRateGenerator::kMaximalCash_ = 20000.0;
 
 double InterestRateGenerator::GetInterestRate(double cash) {
   if (cash == kMinimalCash_) {
